ADD: Added geometry modulo to store data structures (box to crop mask)
